As the heap implementation of our target platforms (WinXP or higher) is good, it's better to use also for the release build new16align / delete16align instead of VirtualAlloc / VirtualFree. Note: in Win9x the heap implementation was quite bad for big and many allocations, virtual function were the preferred choice
git-svn-id: file:///D/data/svnrepos/uimager/trunk@914 15f0fa91-3113-114a-a8b1-9261d5763783

+++ Helpers.cpp
@@ -2066,12 +2066,12 @@ notamd:
     return feature;
 }
 
-LPVOID new16align(unsigned int size)
+LPVOID new16align(SIZE_T size)
 {
 	LPBYTE mem = new BYTE[size+sizeof(LPVOID)+15];
-	LPVOID ptr = (LPVOID)((unsigned int)(mem+sizeof(LPVOID)+15) & ~0x0F);
+	LPVOID ptr = (LPVOID)((SIZE_T)(mem+sizeof(LPVOID)+15) & ~0x0F);
 	((LPVOID*)ptr)[-1] = (LPVOID)mem;
-	ASSERT(((unsigned int)ptr & 0x0F) == 0);
+	ASSERT(((SIZE_T)ptr & 0x0F) == 0);
 	return ptr;
 }
